app/requirements.txt
asgiref==3.7.2
Django==5.0.1
mimetypes-magic==0.4.30
mysqlclient==2.2.4
python-dateutil==2.9.0
setuptools==69.0.3
sqlparse==0.4.4
uWSGI==2.0.23
wheel==0.42.0
pytz==2024.1
PyYAML==6.0.1
django-webpack-loader==0.7.0
djangorestframework==3.15.2
whitenoise
